Add files via upload
Book-My-Show will enable the ads on their website, but they are also very cautious
about their user privacy and information who visit their website. Some ads URL
could contain a malicious link that can trick any recipient and lead to a malware
installation, freezing the system as part of a ransomware attack or revealing
sensitive information. Book-My-Show now wants to analyze that whether the
particular URL is prone to phishing (malicious) or not.